Build Dex image with IPv6 patch
The default Dex image does not support a kubernetes cluster with IPv6.
The cluster network address is not wrapped in square brackets, which
fails when the dex container tries to create custom resources. This
commit builds our own dex image while we are trying to get the fix
upstreamed.

Story: 2006711
Task: 38610

Change-Id: I9193dde6d4260f3c9b5fa12224810f1de0aa53df
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>