Enable helm plugin delivery with the application
This creates a new package spec called python-k8sapp-oidc that will
build two packages: python-k8sapp-oidc and python-k8sapp-oidc-wheels.
These packages will hold all the stevedore plugins needed to support
the application.

These packages are included in the build dependencies for the
stx-oidc-auth-helm package build where the wheels file will
be included in the application tarball.

The helm plugins have been relocated to this repo and provided in
a k8sapp_oidc python module. This module will be extracted from the
wheels and installed on the platform via the sysinv application
framework. The module will be made available when the application
is enabled.

Now that the plugins and the application are residing in the same
repository, GITREVCOUNT will be used to drive the package version of
both to keep them in sync.

Tox test support has been provided for the plugin module.

Change-Id: Iead16c79aa1e3863d8568c9615f4215f2f6ec02b
Story: 2006537
Task: 39108
Depends-On: https://review.opendev.org/713969
Signed-off-by: Angie Wang <angie.wang@windriver.com>
Signed-off-by: Robert Church <robert.church@windriver.com>