Add affinity to dex and oidc-client helm charts
When the standby controller is locked, the oidc pods are scheduled onto
the active controller. When the standby controller is unlocked, the oidc
pods are not scheduled on the standby. This results in four oidc pods on
the active controller.
This update sets the pod anti-affinity rule such that the pod does
not get scheduled onto a node if the node already has the same pod
running.

Closes-Bug: 1865565

Change-Id: I19113ab5a11f0691a0cdba9657138cc3251788cc
Signed-off-by: Teresa Ho <teresa.ho@windriver.com>