Support multiple dex back ends with CLI tool
This commit modifies the CLI tool to support multiple dex backends.
Users with multiple backends configured can select one with "-b".
If only one backend is configured, it will be used if no "-b" is
specified. If multiple backends are configured and "-b" is not
used, the command will fail, and present a list of backends for
the user to pick from.

Story: 2006711
Task: 39696

Change-Id: I532a18a6ec12181a1f3e201374e5e719575cf2fa
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>