oidc-client pod does not properly pick up new config
A config update through helm overrides to oidc-client gets properly
passed to the pod, however, the pod does not restart to pick up the
new config. This commit adds annotations to the oidc-client pod to
track config changes and trigger a pod restart through kubernetes if
the config changes. This is only relevant to pods already deployed on
an applied application. Helm overrides are correctly applied if they
are changed first, then the application applied.

Change-Id: I3d9f48038305283a836827fcebcc2b4e95921147
Closes-Bug: 1923662
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>