Allow secret name to be specified for oidc-client
This commit allows the secret name for the oidc client to be specified
through a helm override. The purpose of this change is to specify the
secret name for dex. Dex helm charts already allow specifying the name
through helm overrides. The same secret is mounted on the client as
well, so this change is required for the entire application to
function. By default, if no overrides are given, the secret name is
assumed to be local-dex.tls, same as the old forced secret name, for
compatibility with older systems.

Change-Id: Id0938c14199d553b786a76633f2f17c94ce16425
Story: 2007361
Task: 42648
Signed-off-by: Jerry Sun <jerry.sun@windriver.com>