Attempt to get zuul and tox passing in this repo
Fix pylint by specifying a nodeset that contains
a python3 interpreter.  Switched the job to run py3.

This application needs to be installed into the virtualenv
in order for its entry points to be discoverable by
stevedore for the unit tests. However py27 fails to find
pbr during the sdist phase since the zuul node is running
a py35 virtualenv.

Disabling sdist should get pip to work. Other tox settings
should install it properly in the virtualenv.

Story: 2008943
Task: 43010
Signed-off-by: albailey <Al.Bailey@windriver.com>
Change-Id: I834a414f90d16d303c499e9ec3799ca5d02ca425