Add taint toleration to oidc-client and dex
A toleration needs to be added to all pods since the
node-role.kubernetes.io/master taint will be restored to all master
nodes. This will ensure that the pod will run on the master node.

The following tests were executed in an AIO-SX setup and lastest STX
dev image:

  - Apply oidc-auth-apps and perform login test. It succeeds.
  - Add taint, apply oidc-auth-apps and perform login test. It fails.
  - Replace oidc-auth-apps with new version (with toleration), add
    taint, apply oidc-auth-apps and perform login test. It succeeds.

Story: 2009232
Task: 43344

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I998599221da52629beed3e7d402efb918b4f55d8