Update debian packages for pkg-versioning
The Debian packaging meta_data file has been changed to reflect all the
latest git commits under the directory, pointed as usable, and to
improve pkg-versioning addressing the first commit as start point to
build packages.

This ensures that any new code submissions under those
directories will increment the versions.

The commit SHA f045d574f3 was chosen to be the BASE_SRCREV of the
python3-oidcauthtools's metadata because is the commit that creates the
debian directory with build  files structure for this package.
(python3-oidcauthtools version updated from .stx.3 to .stx.9)

The commit SHA f045d574f3  was chosen to be the BASE_SRCREV of the
stx-oidc-auth-helm's metadata because is the commit that creates the
debian directory with build files structure for this package. The
commit SHA 638dc8f924 was chosen to be the FILES_GITREVCOUNT of this
package because is the commit that adds the "helm-charts" directory
of the repository referenced in the "src_files".

It is added both GITREVCOUNT and FILES_GITREVCOUNT to calculate the
relevant git commits of "src_path" and "src_files" to this package's
revision.
(stx-oidc-auth-helm version updated from .stx.9 to .stx.36)

The commit SHA 12c005958c was chosen to be the BASE_SRCREV of the
python3-k8sapp-oidc's metadata because is the commit that creates the
debian directory with build files structure for this package.
(python3-k8sapp-oidc version updated from .stx.3 to .stx.11)

Test Plan:
PASS: Verify package versions are updated as expected.
PASS: build-pkgs -c -p python3-oidcauthtools
PASS: build-pkgs -c -p python3-k8sapp-oidc
PASS: build-pkgs -c -p stx-oidc-auth-helm

Story: 2010550
Task: 47498

Signed-off-by: Manoel Benedito Neto <Manoel.BeneditoNeto@windriver.com>
Change-Id: I9f8da9bff27de2c7208df7fd9edf8196d8ab977b