Add default parameters to oidc-auth command
This commit adds default values for parameters "--client" and
"--username" of oidc-auth command.
The value of parameter "--client" is usually the floating IP of the
controller.
The value of parameter "--username" is usually the same as the current
logged in username.

Test Plan:

PASS: In an IPv6 AIO-SX, execute command "oidc-auth -v", type the
password and successfully get the token. Verify in the output that the
value printed for parameter "client" is "https://oamcontroller:30555"
and for parameter "username" is the currently logged in username.
PASS: Repeat the test above in an IPv4 AIO-DX.

Story: 2010738
Task: 48481

Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>
Change-Id: I0c5540066142201044b875f9dd3cc50cc5435fcc