Remove ingress configuration from OIDC client
The ingress configuration of OIDC client was never used and there are
no valid use cases involving it. The implementation contains the
deprecated Kubernetes API "extensions/v1beta1", but instead of fixing
it, the implementation will be removed.

Test Plan:

PASS: Successfully compile a new OIDC app tarball, upload it to an
AIO-SX environment and test OIDC app functionality.

Closes-Bug: 2060055

Change-Id: I22196dc3434113340c73b5918b7a716785fb6463
Signed-off-by: Joao Victor Portal <Joao.VictorPortal@windriver.com>